[COMMON] soc: chipid: add casting for getting 64bit register

+++ drivers/soc/samsung/exynos-chipid.c
@@ -108,7 +108,7 @@ void __init exynos_chipid_early_init(struct device *dev)
 
 	exynos_soc_info.product_id  = __raw_readl(exynos_chipid_base);
 	exynos_soc_info.unique_id  = __raw_readl(exynos_chipid_base + UNIQUE_ID1);
-	exynos_soc_info.unique_id  |= __raw_readl(exynos_chipid_base + UNIQUE_ID2) << 32;
+	exynos_soc_info.unique_id  |= (u64)__raw_readl(exynos_chipid_base + UNIQUE_ID2) << 32;
 	exynos_soc_info.revision = exynos_soc_info.product_id & EXYNOS_REV_MASK;
 }
